requirements.txt
fastapi
starlette
loguru
python-dotenv
google-genai
aiohttp
uvicorn
SQLAlchemy
alembic
psycopg2-binary
redis
python-multipart
faker
email-validator
gunicorn
click
pandas
cryptography
pyjwt
requests
validators
websockets

# python -m pip install --upgrade --no-cache-dir -r requirements.txt